rfm/legacy/requirements.txt
numpy==1.11.0
scikits.audiolab==0.11.0
Pillow==2.7.0
networkx==1.9.1
virtualenv==1.11.6
argparse==1.2.1
pypng==0.0.17
wsgiref==0.1.2
boto==2.34.0
mysql-connector-python==8.0.23
scipy==0.14.0
scikit-learn==0.15.2
joblib==0.8.3
matplotlib==1.4.2
scikits.samplerate==0.3.3
scikit-image==0.11.3
dill==0.2.5
bottle==0.12.10
ws4py==0.3.5
PyJWT==1.4.2
requests==2.13.0